refactor(workflow): split engine.go into focused files
engine.go: 1049→151 lines, structs/interfaces/consts only
engine_dispatch.go: workflow lifecycle + trigger matching
engine_advance.go: step execution loop + transition logic
engine_events.go: agent/human/status event handlers
